Section 26: Everything is done and working perfectly, worked on some slight code refactoring and might do more later.
